Some more reorganization and renaming of files
